Abstracted comments tree logic into dedicated components
